app/requirements.txt
pandas
scikit-learn
lightgbm
joblib
matplotlib
streamlit
google-cloud-storage
google-cloud-bigquery
shap